+++ .changeset/ninety-houses-study.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/wallet-cli": minor
+---
+
+add from and to to swap execute
